Updated unit tests for _check_bgp_neighbor_capability
